Fix various SonarCloud code smells: S1128, S108, S7158, S1905, S1643, S1319, S1172, S2140, S3024, S2184, S6218, S117, S3457, S1151

+++ src/main/java/org/drozdek/graphs/unlam/WeightedDigraph.java
@@ -80,8 +80,8 @@ public class WeightedDigraph {
     }
 
     public boolean createArc(char a1, char a2, int weight) {
-        int n1 = ((int) a1) - 97;
-        int n2 = ((int) a2) - 97;
+        int n1 = a1 - 97;
+        int n2 = a2 - 97;
 
         return createArc(n1, n2, weight);
     }
